Restructure the suffix generation code for output.
The interface.py file contains various plotting functions
which save their output in the output/prfstats/ directory.
However, since we are now saving as per file suffix, we
need to preprocess the path passed as argument to these
functions. We split the path into directory and filename,
and append the run-specific suffix to the directory and
recreate the complete path of the file. This is then used
to create the various directories like heatmap, samples,
weight_functions, etc.