Decommission the misc/ folder.
My understanding of the code is going at a
tragically slow pace, and once I came across utilities.py,
I decided it was a better peice of code to add the run
labeling code instead of having a new directory misc/.
Also, the functionality now allows the user to provide
a customized label instead of labeling randomly.

This will be showcased in the following commit.