Additional tests for new distance metric
These 3 tests work on 3 datasets respectively -- Lorenz,
Pendulum and Rossler. Each dataset directory has multiple
files, each corresponding to an embedding of a scalar
time-series with a specific tau. The test generates PRF's
for each embedding, and then computes the distance from a
reference PRF. It makes a plot of the PRF ref distance as
a function of tau, which can be used to predict the correct
range of tau.